# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAO-BIZ e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  categorySlug: string;
  stock: number;
  isFeatured: boolean;
  isNew: boolean;
  onSale: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  categorySlug: string;
  stock: number;
  isFeatured?: boolean;
  isNew?: boolean;
  onSale?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl?: string;
  categorySlug?: string;
  stock?: number;
  isFeatured?: boolean;
  isNew?: boolean;
  onSale?: boolean;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productImageUrl: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  total: number;
  paymentMethod: string;
  status: string;
  createdAt: string;
}

export interface DeliveryZone {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  isActive: boolean;
  createdAt: string;
}

export interface DeliveryZoneInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  isActive?: boolean;
}

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  paymentMethod: string;
  /** @nullable */
  deliveryZoneId?: number | null;
  /** @nullable */
  deliveryZoneName?: string | null;
  /** @nullable */
  deliveryFee?: number | null;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface AdminStats {
  todayOrders: number;
  todayRevenue: number;
  totalProductsSold: number;
  lowStockCount: number;
  pendingOrders: number;
  confirmedOrders: number;
  deliveredOrders: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
onSale?: boolean;
isNew?: boolean;
};

export type ListOrdersParams = {
status?: string;
};