# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAO-BIZ e-commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product management
  - name: categories
    description: Product categories
  - name: orders
    description: Order management
  - name: admin
    description: Admin operations
  - name: delivery-zones
    description: Delivery zone management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: onSale
          in: query
          required: false
          schema:
            type: boolean
        - name: isNew
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard stats (admin)
      responses:
        "200":
          description: Admin dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /delivery-zones:
    get:
      operationId: listDeliveryZones
      tags: [delivery-zones]
      summary: List active delivery zones (public)
      responses:
        "200":
          description: List of active delivery zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryZone"

  /admin/delivery-zones:
    get:
      operationId: adminListDeliveryZones
      tags: [admin]
      summary: List all delivery zones (admin)
      responses:
        "200":
          description: All delivery zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryZone"
    post:
      operationId: createDeliveryZone
      tags: [admin]
      summary: Create a delivery zone (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryZoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryZone"

  /admin/delivery-zones/{id}:
    put:
      operationId: updateDeliveryZone
      tags: [admin]
      summary: Update a delivery zone (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryZoneInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryZone"
    delete:
      operationId: deleteDeliveryZone
      tags: [admin]
      summary: Delete a delivery zone (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
      required:
        - id
        - name
        - slug
        - icon

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        categorySlug:
          type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        onSale:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - categorySlug
        - stock
        - isFeatured
        - isNew
        - onSale
        - createdAt

    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        categorySlug:
          type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        onSale:
          type: boolean
      required:
        - name
        - description
        - price
        - imageUrl
        - categorySlug
        - stock

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        categorySlug:
          type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        onSale:
          type: boolean

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - productId
        - productName
        - productImageUrl
        - quantity
        - unitPrice

    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerPhone
        - customerAddress
        - items
        - total
        - paymentMethod
        - status
        - createdAt

    DeliveryZone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - isActive
        - createdAt

    DeliveryZoneInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        isActive:
          type: boolean
      required:
        - name
        - price

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        paymentMethod:
          type: string
        deliveryZoneId:
          type: ["integer", "null"]
        deliveryZoneName:
          type: ["string", "null"]
        deliveryFee:
          type: ["number", "null"]
      required:
        - customerName
        - customerPhone
        - customerAddress
        - items
        - paymentMethod

    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AdminStats:
      type: object
      properties:
        todayOrders:
          type: integer
        todayRevenue:
          type: number
        totalProductsSold:
          type: integer
        lowStockCount:
          type: integer
        pendingOrders:
          type: integer
        confirmedOrders:
          type: integer
        deliveredOrders:
          type: integer
      required:
        - todayOrders
        - todayRevenue
        - totalProductsSold
        - lowStockCount
        - pendingOrders
        - confirmedOrders
        - deliveredOrders