# lib/api-zod/src/generated/types/deliveryZoneInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAO-BIZ e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryZoneInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  isActive?: boolean;
}