# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAO-BIZ e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  paymentMethod: string;
  /** @nullable */
  deliveryZoneId?: number | null;
  /** @nullable */
  deliveryZoneName?: string | null;
  /** @nullable */
  deliveryFee?: number | null;
}