# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAO-BIZ e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl?: string;
  categorySlug?: string;
  stock?: number;
  isFeatured?: boolean;
  isNew?: boolean;
  onSale?: boolean;
}